+++ app/src/main/java/com/killjoy/stuntion/features/presentation/utils/components/Wallet.kt
@@ -8,7 +8,6 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.RoundedCornerShape
@@ -27,7 +26,7 @@ import com.killjoy.stuntion.ui.theme.Type
 @Composable
 fun Wallet(
     modifier: Modifier = Modifier,
-    balance: Double,
+    balance: Int,
     onTopUpClicked: () -> Unit,
 ) {
     Card(
